fixed errors

+++ internal/api/post_input.go
@@ -3,5 +3,5 @@ package api
 type PostInput struct {
 	Author int    `json:"author"`
 	Body   string `json:"body"`
-	Likes  int    `json:"likes"`
+	Likes  int64  `json:"likes"`
 }
